+++ ImplementationPhase/Model/NeuralNetworkModelModule/include/OutputStorageLayer.hpp
@@ -8,14 +8,14 @@
 class OutputStorageLayer  : public NetworkLayer {
 
   private:
-    MATRIX_3D(double) outputData;
+    MATRIX_3D(float) outputData;
 
   public:
     OutputStorageLayer();
-    OutputStorageLayer(std::string name, MATRIX_3D(double) outputData);
-    OutputStorageLayer(std::string name, MATRIX_3D(double) outputData, int* inputDimensions);
-    void setData(MATRIX_3D(double) outputData);
-    MATRIX_3D(double) getData();
+    OutputStorageLayer(std::string name, MATRIX_3D(float) outputData);
+    OutputStorageLayer(std::string name, MATRIX_3D(float) outputData, int* inputDimensions);
+    void setData(MATRIX_3D(float) outputData);
+    MATRIX_3D(float) getData();
     // ~CollectResultsLayer();
 };
 #endif
